Move preload into one file, (hopefully temporary)
I can do it in multiple files when hooking to the inter-comm channels, but the issue occurs when calling those async functions in the render scripts.

This basically means everything would have to wait for the preloads to run, so probably best to just keep everything confined to one file.